refactor(extraction): split grounding helpers
